코드 분리 및 폴더 정리
