+++ DrawingApp/DrawingApp/Protocol/EventDelegate.swift
