+++ DrawingApp/DrawingApp/Protocol/PositionProtocol.swift
@@ -1,0 +1,10 @@
+//
+//  PositionProtocol.swift
+//  DrawingApp
+//
+//  Created by elly on 11/20/23.
+//
+
+import Foundation
+
+protocol PositionProtocol: Identifiable, Pointable { }
